Revert "Refactor: Move Prediction and Poll handling in their own file (#5)"
This reverts commit dad25d063bf467005d8e07ac8c43f08e590d64b4.
